requirements.txt
Django==2.2.28
django-braces==1.13.0
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
django-environ==0.4.5
django-jet==1.0.8
django-sendsms==0.3.1
djangorestframework==3.11.2
djangorestframework-jwt==1.11.0
drf-user==0.0.8
drf-yasg==1.17.1
drfaddons==0.1.4
huey==2.4.4
pre-commit==2.20.0
psycopg2-binary==2.8.6
python-dateutil==2.8.2
redis==4.4.2
sentry-sdk==1.14.0
six==1.16.0
